refactor(solver): split function-failure explanation out of explain.rs
`subtype/explain.rs` exceeded the repository-wide 2000-line hard limit (§19)
after the union-source elaboration change. Move the function/callable
signature failure-explanation methods (`explain_function_failure`,
`explain_function_to_callable_failure`,
`explain_callable_to_callable_signature_failure`, and their helpers) into a
new `subtype/explain_function.rs` shard. Pure code movement; the entry points
are `pub(super)` so the main `explain_failure_inner` dispatch still calls them.
explain.rs is now 1775 lines.
